# Internally avoid element verification cost of `.view(GF)`