--- a/README.md
+++ b/README.md
@@ -221,7 +221,7 @@ This library would not be possible without all of the other libraries mentioned.
 
 If this library was useful to you in your research, please cite us. Following the [GitHub citation standards](https://docs.github.com/en/github/creating-cloning-and-archiving-repositories/creating-a-repository-on-github/about-citation-files), here is the recommended citation.
 
-### Bibtex
+### BibTeX
 
 ```bibtex
 @software{Hostetter_Galois_2020,